+++ src/components/NavBar/GameOptionsModal/GameOptionsModal.vue
@@ -15,14 +15,20 @@
                 </div>
                 <div class="modal-body h-100 visible-scrollbar">
                     <ul id="game-summary-modal-tabs" class="nav nav-pills nav-fill">
-                        <li id="game-results-tab" class="nav-item" @click="openGameRolesOptions">
+                        <li id="game-roles-options-tab" class="nav-item" @click="openGameRolesOptions">
                             <a class="nav-link" :class="{ active: panel === 'game-roles-options' }" href="#">
                                 <RequiredActionIcon v-if="!game.areGameRolesOptionsValid" class="mr-2" color="white"/>
                                 <i class="fa fa-chess mr-2"/>
                                 <span v-html="$t('GameOptionsModal.gameRolesOptions')"/>
                             </a>
                         </li>
-                        <li id="game-history-tab" class="nav-item" @click="openGameRepartitionOptions">
+                        <li id="game-players-position-tab" class="nav-item" @click="openGamePlayersPosition">
+                            <a class="nav-link" :class="{ active: panel === 'game-players-position' }" href="#">
+                                <i class="fa fa-users mr-2"/>
+                                <span v-html="$t('GameOptionsModal.gamePlayersPosition')"/>
+                            </a>
+                        </li>
+                        <li id="game-repartition-options-tab" class="nav-item" @click="openGameRepartitionOptions">
                             <a class="nav-link" :class="{ active: panel === 'game-repartition-options' }" href="#">
                                 <i class="fa fa-random mr-2"/>
                                 <span v-html="$t('GameOptionsModal.gameRepartitionOptions')"/>
@@ -37,6 +43,8 @@
                                           @options-updated="optionsUpdated = true"/>
                         <GameRepartitionOptions v-else-if="panel === 'game-repartition-options'" key="game-repartition-options"
                                                 @options-updated="optionsUpdated = true"/>
+                        <GamePlayersPosition v-else-if="panel === 'game-players-position'" key="game-repartition-options"
+                                             @options-updated="optionsUpdated = true"/>
                     </transition>
                 </div>
                 <div class="modal-footer">
@@ -54,14 +62,15 @@ import GameRolesOptions from "@/components/NavBar/GameOptionsModal/GameRolesOpti
 import GameRepartitionOptions from "@/components/NavBar/GameOptionsModal/GameRepartitionOptions/GameRepartitionOptions";
 import RequiredActionIcon from "@/components/shared/RequiredActionIcon";
 import { mapGetters } from "vuex";
+import GamePlayersPosition from "@/components/NavBar/GameOptionsModal/GamePlayersPosition";
 
 export default {
     name: "GameOptionsModal",
-    components: { RequiredActionIcon, GameRepartitionOptions, GameRolesOptions },
+    components: { GamePlayersPosition, RequiredActionIcon, GameRepartitionOptions, GameRolesOptions },
     data() {
         return {
             optionsUpdated: false,
-            panel: "c",
+            panel: "game-roles-options",
         };
     },
     computed: { ...mapGetters("game", { game: "game" }) },
@@ -85,6 +94,9 @@ export default {
         },
         openGameRolesOptions() {
             this.panel = "game-roles-options";
+        },
+        openGamePlayersPosition() {
+            this.panel = "game-players-position";
         },
         openGameRepartitionOptions() {
             this.panel = "game-repartition-options";